synaptics_3k: Do not hardcode pressure & width
The synaptics_3k driver has the pressure and width values hardcoded
into the params. This makes adjusting for width and pressure slightly
more confusing/work than it should be. Get rid of the hardcoded values
and use defines.

[NOTE: Using MAX_WIDTH for ABS_MT_PRESSURE may not be correct, but we'll
keep it that way for now, since that's the way the driver seems to be
configured.]

idea is taken here -
https://github.com/poondog/joey-m7-GPE/commit/d2d9cb23d2eb1ad3ce36084ecdaf08042ad5ff19