writeback: I believe these defines were calculated with HZ=1000 in mind
Since our devices are HZ=100 I hardcoded the values on this constants.
I'm not 100% sure if this code was made for HZ with 1000 but I couldn't
track it to understand clearly. Although I didn't see any drawback of
using this as it is now. If anyone knows the answer please chime in.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

Conflicts:

mm/page-writeback.c

idea is taken here -
https://github.com/zarboz/M7wlv_KK_kernel/commit/8b5971f8b9fc698457044ffeaa5cacd8f7eae1dd