mmc: msm_sdcc: enable async suspend
This is aimed to reduce the suspend time of msm_sdcc device.
The suspend
of this device would be carried out in a separate
thread and other
devices may suspend concurrently. This is
required since the mmc
sub-system contributes to the major
chunk of suspend latency.

idea is
taken here -
https://github.com/franciscofranco/hammerhead/commit/5d44bde0f927e012eda152824a2e57c2081bce17