msm: peripheral-loader: Change IOMAP_SIZE from SZ_4M to SZ_1M
The ioremap call made for zeroing out trailing memory is
currently
requesting for a 4M memory region. This is resulting in
vmalloc
allocation failures on some platforms.
Instead, request for 1M
memory locations and zero out the area
in smaller chunks to make this
failure less probable.

Signed-off-by: Deepak Katragadda
<dkatraga@codeaurora.org>
Signed-off-by: Pranav Vashi
<neobuddy89@gmail.com>
Signed-off-by: franciscofranco
<franciscofranco.1990@gmail.com>

idea is taken here -
https://github.com/franciscofranco/hammerhead/commit/1523114b66319a638f8efccfa5046d3211636164