USB: Increase reset timeout.
John's NEC 0.96 xHCI host controller needs a longer timeout for a
warm
reset to complete.  The logs show it takes 650ms to complete the
warm
reset, so extend the hub reset timeout to 800ms to be on the safe
side.

This commit should be backported to kernels as old as 3.2, that
contain
the commit 75d7cf7 "usbcore: refine warm reset
logic".

Signed-off-by: Sarah Sharp
<sarah.a.sharp@linux.intel.com>
Acked-by: Alan Stern
<stern@rowland.harvard.edu>
Reported-by: John Covici
<covici@ccs.covici.com>
Signed-off-by: Greg Kroah-Hartman
<gregkh@linuxfoundation.org>

idea is taken here -
https://github.com/varunchitre15/thunderzap_xl/commit/d5a174583f4194941c2a8672c558ebfa1febf330