USB: hub: handle claim of enabled remote wakeup after reset
Some touchscreens have buggy firmware which claims
remote wakeup to be
enabled after a reset. They nevertheless
crash if the feature is cleared
by the host.
Add a check for reset resume before checking for
an enabled
remote wakeup feature. On compliant
devices the feature must be cleared
after a reset anyway.

Signed-off-by: Oliver Neukum
<oneukum@suse.de>
Acked-by: Alan Stern
<stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman
<gregkh@linuxfoundation.org>

idea is taken here -
https://github.com/varunchitre15/thunderzap_xl/commit/ca0f8c8acbba855cf08de5480062795d9631d1cd