USB: Don't use EHCI port sempahore for USB 3.0 hubs.
The EHCI host controller needs to prevent EHCI initialization when
the
UHCI or OHCI companion controller is in the middle of a port reset.
It
uses ehci_cf_port_reset_rwsem to do this.  USB 3.0 hubs can't be
under
an EHCI host controller, so it makes no sense to down the
semaphore for
USB 3.0 hubs.  It also makes the warm port reset code more
complex.

Don't down ehci_cf_port_reset_rwsem for USB 3.0
hubs.

Signed-off-by: Sarah Sharp
<sarah.a.sharp@linux.intel.com>
Acked-by: Alan Stern
<stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman
<gregkh@linuxfoundation.org>

idea is taken here -
https://github.com/varunchitre15/thunderzap_xl/commit/605d4dce7d3b8a458e240e28e309b63e503e1a71