usbfs: Always allow ctrl requests with USB_RECIP_ENDPOINT on the ctrl ep
When usbfs receives a ctrl-request from userspace it calls
check_ctrlrecip,
which for a request with USB_RECIP_ENDPOINT tries to
map this to an interface
to see if this interface is claimed, except for
ctrl-requests with a type of
USB_TYPE_VENDOR.

When trying to use this
device: http://www.akaipro.com/eiepro
redirected to a Windows vm running
on qemu on top of Linux.

The windows driver makes a ctrl-req with
USB_TYPE_CLASS and
USB_RECIP_ENDPOINT with index 0, and the mapping of
the endpoint (0) to
the interface fails since ep 0 is the ctrl endpoint
and thus never is
part of an interface.

This patch fixes this ctrl-req
failing by skipping the checkintf call for
USB_RECIP_ENDPOINT ctrl-reqs
on the ctrl endpoint.

Reported-by: Dave Stikkolorum
<d.r.stikkolorum@hhs.nl>
Tested-by: Dave Stikkolorum
<d.r.stikkolorum@hhs.nl>
Signed-off-by: Hans de Goede
<hdegoede@redhat.com>
Acked-by: Alan Stern
<stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman
<gregkh@linuxfoundation.org>

idea is taken here -
https://github.com/varunchitre15/thunderzap_xl/commit/69642fb9cd0fd44d02303659ba3d89b4cb9ac36a