usb: core: get config and string descriptors for unauthorized devices
There is no need to skip querying the config and string descriptors
for
unauthorized WUSB devices when usb_new_device is called.  It is
allowed
by WUSB spec.  The only action that needs to be delayed
until
authorization time is the set config.  This change allows user
mode
tools to see the config and string descriptors earlier in
enumeration
which is needed for some WUSB devices to function properly
on Android
systems.  It also reduces the amount of divergent code paths
needed
for WUSB devices.

Signed-off-by: Thomas Pugliese
<thomas.pugliese@gmail.com>
Signed-off-by: Greg Kroah-Hartman
<gregkh@linuxfoundation.org>

idea is taken here -
https://github.com/varunchitre15/thunderzap_xl/commit/ee161026342fd713232ef053ac4c5cea0084ae35