ARM: dts: msm: lower LDO 11 minimum voltage constraint for msm8974
Change the minimum voltage constraint of PM8941 LDO 11 from 1.3 V
to
1.225 V.  This allows the RPM to reduce the voltage of LDO 11
when no
consumers on the application processor require it to be
at the maximum
level of 1.3 V.  This is particularly important
when a consumer on
another processor needs LDO 11 to be enabled
but no consumer on the
application processor requires the same.

Make the same change for
PMA8084 LDO 11 on msm8974pro boards.

Change-Id:
I9cf71cf7d9360c045af1ac51a3994e3f6d672191
CRs-Fixed:
614000
Signed-off-by: David Collins
<collinsd@codeaurora.org>
Signed-off-by: flar2
<asegaert@gmail.com>

idea is taken here -
https://github.com/flar2/ElementalX-N5/commit/1235feb82e930dd29c060e9df69a36b3f88cde37