futex: Prevent attaching to kernel threads
We happily allow userspace to declare a random kernel thread to be
the
owner of a user space PI futex.

Found while analysing the fallout
of Dave Jones syscall fuzzer.

We also should validate the thread group
for private futexes and find
some fast way to validate whether the
"alleged" owner has RW access on
the file which backs the SHM, but
that's a separate issue.

Change-Id:
I7444c58d00cbded35b078331a8740e8431cb13a5

idea is taken here -
https://github.com/rmbq/android_kernel_htc_msm8960/commit/a6e2cf95f013e667262352176806a43aee817217